requirements.txt
django
djangorestframework
drf-yasg
drf-yasg[validation]
